13.06

+++ RealEstate.cs
@@ -13,7 +13,7 @@ namespace RealEstateProject
     {
         public string Type { get; set; }
         public double Price { get; set; }
-        public float Surface { get; set; }
+        public double Surface { get; set; }
         public string City { get; set; }
         public double Rent { get; set; }
         public string Market { get; set; }
@@ -21,7 +21,7 @@ namespace RealEstateProject
         public static int RealEstateCount = 1;
 
         //działka
-        public RealEstate(double price, float surface, Cities city, Markets market)
+        public RealEstate(double price, double surface, Cities city, Markets market)
         {
             this.Type = Types.Działka.ToString();
             this.Price = price;
@@ -38,7 +38,7 @@ namespace RealEstateProject
         }
 
         //mieszkanie
-        public RealEstate(double price, float surface, Cities city, double rent, Markets market)
+        public RealEstate(double price, double surface, Cities city, double rent, Markets market)
         {
             this.Type = Types.Mieszkanie.ToString();
             this.Price = price;
